requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0
scikit-learn>=1.0.0
pyyaml>=6.0
tqdm>=4.62.0
pytest>=6.2.0
jupyter>=1.0.0
plotly>=5.0.0
dash>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
